requirements.txt
numpy==1.26.1
pandas==1.5.3
plotly==5.16.1
kaleido==0.2.1
PyYAML==6.0.1
scikit_learn==1.3.1
torch==2.0.1
tqdm==4.65.0
